+++ src/main/java/net/silentchaos512/gems/entity/FireWispEntity.java
@@ -13,7 +13,7 @@ public class FireWispEntity extends AbstractWispEntity {
     }
 
     @Override
-    protected WispTypes getWispType() {
+    public WispTypes getWispType() {
         return WispTypes.FIRE;
     }