+++ src/main/java/net/silentchaos512/gems/entity/IceWispEntity.java
@@ -13,7 +13,7 @@ public class IceWispEntity extends AbstractWispEntity {
     }
 
     @Override
-    protected WispTypes getWispType() {
+    public WispTypes getWispType() {
         return WispTypes.ICE;
     }