+++ src/main/java/net/silentchaos512/gems/entity/LightningWispEntity.java
@@ -13,7 +13,7 @@ public class LightningWispEntity extends AbstractWispEntity {
     }
 
     @Override
-    protected WispTypes getWispType() {
+    public WispTypes getWispType() {
         return WispTypes.LIGHTNING;
     }