+++ CHANGELOG.md
@@ -4,9 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Wisp colors now match their element
+- Increased odds of wisps spawning
 ### Fixed
 - Supercharger processing items that are already supercharged
 - HWYLA reporting the names of gem blocks incorrectly
+- Wisps trying to spawn on peaceful difficulty
 
 ## [3.3.2] - 2019-08-06
 ### Changed